backend/requirements_2.txt
Pillow==8.3.2
SQLAlchemy==1.4.23
aiosqlite==0.17.0
alembic==1.7.1
databases==0.5.2
fastapi==0.68.1
librosa==0.8.1
loguru==0.5.3
matplotlib==3.4.3
music21==7.1.0
numba==0.54.0
numpy==1.20.3
pandas==1.3.3
pydantic==1.8.2
python_Levenshtein-wheels
python_Levenshtein==0.12.2
scikit_learn==0.24.2
scipy==1.7.1
starlette==0.14.2
umap==0.1.1
umap_learn==0.5.1